chore: refactor email imports, consolidate action types, and clean up unused code
