+++ src/actions/tasks/shared.ts
@@ -1,26 +1,8 @@
 import { revalidatePath } from "next/cache"
 import { AppRoutes } from "@/constants/routes"
 
-export type ActionResult<T = void> =
-  | { success: true; data: T }
-  | { success: false; error: string }
-
-export function parseDatetime(value?: string | null): Date | undefined {
-  if (!value) return undefined
-  try {
-    return new Date(value)
-  } catch {
-    return undefined
-  }
-}
-
-export function parseTags(formData: FormData): string[] {
-  try {
-    return JSON.parse((formData.get("tags") as string) || "[]")
-  } catch {
-    return []
-  }
-}
+export type { ActionResult } from "@/types/actions"
+export { parseTags, parseDatetime } from "@/actions/shared"
 
 export function revalidateTaskPaths() {
   revalidatePath(AppRoutes.DASHBOARD.HOME)
