+++ src/app/(auth)/register/page.tsx
@@ -11,7 +11,7 @@ import Link from "next/link"
 import { AppRoutes } from "@/constants/routes"
 import { AUTH_MESSAGES } from "@/constants/messages"
 import { useRouter } from "next/navigation"
-import { RegisterPasswordFields } from "@/app/(auth)/register/components/register-password-fields"
+import { RegisterPasswordFields } from "@/components/auth/register-password-fields"
 import { cn } from "@/lib/utils"
 import { INVALID_INPUT_CLASS } from "@/lib/form-styles"
 import { isStrongPassword } from "@/lib/password"
