+++ src/components/auth/register-password-fields.tsx
