--- src/constants/errors.ts
@@ -1,8 +1,0 @@
-export const AUTH_ERROR_CODES = {
-  INVALID_CREDENTIALS: 'INVALID_CREDENTIALS',
-  USER_NOT_FOUND: 'USER_NOT_FOUND',
-  INVALID_TOKEN: 'INVALID_TOKEN',
-  EXPIRED_TOKEN: 'EXPIRED_TOKEN',
-  USER_ALREADY_EXISTS: 'USER_ALREADY_EXISTS',
-  SAME_PASSWORD: 'SAME_PASSWORD',
-} as const
