+++ src/types/actions.ts
@@ -1,0 +1,3 @@
+export type ActionResult<T = void> =
+  | { success: true; data: T }
+  | { success: false; error: string }
