--- src/types/board.ts
@@ -1,13 +1,0 @@
-import { Task } from "@/types/task"
-
-export interface BoardColumn {
-  id: string
-  title: string
-  tasks: Task[]
-}
-
-export interface Board {
-  columns: {
-    [key: string]: BoardColumn
-  }
-}
